Add selection box (#5)
This adds the ability to select points using a sub-rectangle in the canvas.

That rectangle is un-projected from the canvas to form a frustum in world coordinates. Points inside the frustum are selected. This is mostly adapted from the related three.js examples with appropriate links and some modifications.

To select points you need to toggle between camera/select mode, which determines what mouse movements do in the canvas. I also made holding shift enable selection mode as a quick and natural way to use this.

Selecting points currently changes their color to white. The selection can only be reset by changing the time point.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>