Refactor selection state and selectors (#88)
This refactors the current selection so that it is state on `PointCanvas` instead of react state, using the existing reducer.

This also refactors the box selection mechanism so that it shares a similar interface to the spherical selectors. `PointSelector` now owns both of these and delegates appropriately based on the selection mode.

I think this also fixes some visibility toggle bugs with the spherical selectors (e.g. on `main` i think that using `s` to toggle the visibility of `SPHERICAL_SELECTOR` can cause the `SPHERE` controls to be shown) and hopefully makes them stay in sync more (I think they're supposed to do that?).

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>